requirements.txt
fastapi
uvicorn[standard]
tensorflow
pandas
google-cloud-firestore
google-cloud-storage
python-dotenv
